Simplified the API.

+++ v7/module_test.go
@@ -3109,7 +3109,7 @@ func TestBinary(t *tes.T) {
 	ass.Equal(t, byte(0xf8), v.GetValue(-1))
 	ass.Equal(t, v.AsArray(), fra.Binary(v.AsArray()).AsArray())
 	ass.Equal(t, b2, fra.BinaryFromSequence(v.GetValues(1, 3)).AsString())
-	ass.Equal(t, uint(1), v.GetIndex(0x69))
+	ass.Equal(t, 1, v.GetIndex(0x69))
 }
 
 func TestBinaryLibrary(t *tes.T) {
@@ -3154,7 +3154,7 @@ func TestName(t *tes.T) {
 	var v2 = fra.Name(v1.AsArray())
 	ass.Equal(t, v1.AsString(), v2.AsString())
 	var v3 = fra.NameFromSequence(v1.GetValues(1, 2))
-	ass.Equal(t, 1, int(v1.GetIndex("bali-nebula")))
+	ass.Equal(t, 1, v1.GetIndex("bali-nebula"))
 	ass.Equal(t, "/bali-nebula/types", v3.AsString())
 }
 
@@ -3299,7 +3299,7 @@ func TestQuote(t *tes.T) {
 	ass.Equal(t, 'a', rune(v.GetValue(1)))
 	ass.Equal(t, '4', rune(v.GetValue(-1)))
 	ass.Equal(t, `"d本1"`, fra.QuoteFromSequence(v.GetValues(4, 6)).AsString())
-	ass.Equal(t, 8, int(v.GetIndex('3')))
+	ass.Equal(t, 8, v.GetIndex('3'))
 }
 
 func TestQuotesLibrary(t *tes.T) {
@@ -3330,7 +3330,7 @@ func TestVersion(t *tes.T) {
 	ass.Equal(t, uint(1), v1.GetValue(1))
 	ass.Equal(t, uint(3), v1.GetValue(-1))
 	var v3 = fra.VersionFromSequence(v1.GetValues(1, 2))
-	ass.Equal(t, 2, int(v1.GetIndex(2)))
+	ass.Equal(t, 2, v1.GetIndex(2))
 	ass.Equal(t, "v1.2", v3.AsString())
 }
